tunnig api responses / trying to migration implementates with sqlite

+++ api/app/routes/clients/delete.php
@@ -4,15 +4,27 @@ use lalocespedes\Models\Clients;
 
 $app->delete('/clients/:id', function($id) use($app) {
 
-  $client = Clients::find($id);
+    $client = Clients::find($id);
 
-  $client->delete();
+    if($client) {
 
-  $return = [
+      $client->delete();
 
-    'message' => 'Cliente borrado'
+      $return = [
 
-  ];
+        'message' => 'Cliente borrado'
+
+      ];
+
+    } else {
+
+      $return = [
+
+        'message' => 'Registro no encontrado'
+
+      ];
+
+    }
 
   $response = $app->response();
   $response->header('Access-Control-Allow-Origin', '*');
